# fix(schematics): speed up the dry-run command