--- a/packages/schematics/src/collection/ngrx/index.ts
+++ b/packages/schematics/src/collection/ngrx/index.ts
@@ -25,6 +25,7 @@ import {
 } from './rules';
 import { formatFiles } from '../../utils/rules/format-files';
 import { NodePackageInstallTask } from '@angular-devkit/schematics/tasks';
+import { excludeUnnecessaryFiles } from '@nrwl/schematics/src/utils/rules/filter-tree';
 
 /**
  * Rule to generate the Nx 'ngrx' Collection
@@ -88,6 +89,7 @@ function generateNgrxFilesFromTemplates(options: Schema) {
 
   const templateSource = apply(url('./files'), [
     !options.facade ? filter(excludeFacade) : noop(),
+    excludeUnnecessaryFiles(),
     template({ ...options, tmpl: '', ...names(name) }),
     move(moduleDir)
   ]);