--- a/packages/schematics/src/collection/workspace-schematic/index.ts
+++ b/packages/schematics/src/collection/workspace-schematic/index.ts
@@ -11,10 +11,12 @@ import {
 import { Schema } from './schema';
 import { toFileName } from '../../utils/name-utils';
 import { formatFiles } from '../../utils/rules/format-files';
+import { excludeUnnecessaryFiles } from '@nrwl/schematics/src/utils/rules/filter-tree';
 
 export default function(schema: Schema): Rule {
   const options = normalizeOptions(schema);
   const templateSource = apply(url('./files'), [
+    excludeUnnecessaryFiles(),
     template({
       dot: '.',
       tmpl: '',