--- a/packages/schematics/src/utils/rules/filter-tree.ts
+++ b/packages/schematics/src/utils/rules/filter-tree.ts
@@ -1,0 +1,15 @@
+import { filter, Rule } from '@angular-devkit/schematics';
+
+/**
+ * Exclude common paths from the working Tree
+ */
+export function excludeUnnecessaryFiles(): Rule {
+  return filter(
+    path =>
+      !path.startsWith('/node_modules') &&
+      !path.startsWith('/dist') &&
+      !path.startsWith('/.git') &&
+      !path.startsWith('/.vscode') &&
+      !path.startsWith('/.idea')
+  );
+}