## Description

Currently the `dry-run` command speed is on an average of:
* `ng g app app_name --dry-run`: 7099.051ms to 7207.579ms;
* `ng g lib lib_name --dry-run`: 4015.622ms to 3881.893ms;

This is due to too many files in the tree. While performing the transformation required by the schematic, the tree includes files in the *node_modules, .git, .idea, .vscode* folders.
These files are not needed and impact the performance of the commandline.

This add a filter function on the tree, before anything is moved. With that filter function
the average speed is:
* `ng g app app_name --dry-run`: 1420.826ms to 1435.487ms;
* `ng g lib lib_name --dry-run`: 1382.279ms to 1413.042ms;

## Issue
close #706